Implement user balance functionality
- Add withdrawals table migration - Update BalanceResponse model to use
float64 for withdrawn - Implement balance calculation in repository -
Add balance endpoint handler with auth check - Expose balance methods
through service layer

+++ internal/models/http.go
@@ -21,5 +21,5 @@ type BalanceWithdrawRequest struct {
 // GET /api/user/balance HTTP/1.1
 type BalanceResponse struct {
 	Current   float64 `json:"current"`
-	Withdrawn int64   `json:"withdrawn"`
+	Withdrawn float64 `json:"withdrawn"`
 }
